+++ src/bp/core/server.ts
@@ -47,7 +47,6 @@ import { MonitoringService } from './services/monitoring'
 import { NotificationsService } from './services/notification/service'
 import { WorkspaceService } from './services/workspace-service'
 import { TYPES } from './types'
-import RealtimeService from './services/realtime'
 
 const BASE_API_PATH = '/api/v1'
 const SERVER_USER_STRATEGY = 'default' // The strategy isn't validated for the userver user, it could be anything.
@@ -103,8 +102,7 @@ export default class HTTPServer {
     @inject(TYPES.BotService) private botService: BotService,
     @inject(TYPES.AuthStrategies) private authStrategies: AuthStrategies,
     @inject(TYPES.MonitoringService) private monitoringService: MonitoringService,
-    @inject(TYPES.AlertingService) private alertingService: AlertingService,
-    @inject(TYPES.RealtimeService) private realtime: RealtimeService
+    @inject(TYPES.AlertingService) private alertingService: AlertingService
   ) {
     this.app = express()
 
@@ -159,8 +157,7 @@ export default class HTTPServer {
       authService,
       ghostService,
       workspaceService,
-      logger: this.logger,
-      realtime
+      logger: this.logger
     })
   }
