--- src/bp/core/services/dialog/flow/typings.ts
@@ -1,8 +1,0 @@
-type Modification = 'rename' | 'delete' | 'create' | 'update'
-
-export interface FlowModification {
-  name: string
-  modification: Modification
-  newName?: string
-  payload?: any
-}
